requirements.txt
beautifulsoup4==4.12.3
firebase-admin==6.5.0
google-cloud-storage==2.18.2
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.42.2
pytest==8.3.4
